refactor: 상세정보 api 요청 Detail컴포넌트로 변경 #62
